+++ src/app/components/forms/form/form-container/terra-form-container.component.ts
@@ -15,6 +15,7 @@ import {
 import { TerraFormFieldInterface } from '../model/terra-form-field.interface';
 import { TerraKeyValuePairInterface } from '../../../../models/terra-key-value-pair.interface';
 import {
+    AbstractControl,
     ControlValueAccessor,
     FormGroup,
     NG_VALUE_ACCESSOR
@@ -124,6 +125,25 @@ export class TerraFormContainerComponent implements OnInit, OnChanges, ControlVa
             else
             {
                 this.formFieldVisibility[field.key] = isNullOrUndefined(field.value.isVisible) || field.value.isVisible;
+            }
+
+            if(!isNullOrUndefined(this.formGroup))
+            {
+                let control:AbstractControl = this.formGroup.get(field.key);
+                if(this.formFieldVisibility[field.key])
+                {
+                    if(control.disabled)
+                    {
+                        control.enable({onlySelf:true});
+                    }
+                }
+                else
+                {
+                    if(control.enabled)
+                    {
+                        control.disable({onlySelf:true});
+                    }
+                }
             }
         });
     }