+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 * **terra-form** no longer validates hidden form fields
 
 ### Features
-* **terra-form.helper** instead of the usual required validator, the requiredTrue validator is now added for controls of type checkbox.
 * **tc-radio-group** allow the radio group to accept more elements than only the tc-radio-input. It helps to create more structured combinations of Elements.  
+* **terra-card** added possibility to overwrite the background color of the footer
+* **terra-form.helper** instead of the usual required validator, the requiredTrue validator is now added for controls of type checkbox and toggle.
 
 <a name="3.9.0"></a>
 # 3.9.0 (10.04.2019)